Included tests for boolean and time filters.
